release: 发布v3.2.9

+++ include/motor/types.h
@@ -343,7 +343,7 @@ struct MotorParameterTraits<MotorParameter::PosKpKd> {
 
 template <>
 struct MotorParameterTraits<MotorParameter::BrakeStatus> {
-    using RawType = uint16_t;
+    using RawType = uint8_t;
     using RetType = uint16_t;
 };
